Update project architecture
